Docket_management_system/Docket-system-backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
python-dotenv
passlib
PyJWT
gunicorn
qrcode
reportlab
bcrypt==3.2.0
